// src/api/nexuscore/model/questOut.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Nexuscore
 * The backend for all Everthorn internal services
 * OpenAPI spec version: 1.0.0
 */
import type { ObjectiveOut } from './objectiveOut';
import type { UserOut } from './userOut';

export interface QuestOut {
  /** The Quest ID */
  quest_id: number;
  /** The time that this quest begins to be able to be accepted */
  start_time: string;
  /** The time that this quest will no longer be available to be accepted */
  end_time: string;
  /** The quest title */
  title: string;
  /** The quest description */
  description: string;
  /** A list of tags describing this quest */
  tags: string[];
  /** The quest type */
  quest_type: string;
  created_by: UserOut;
  objectives: ObjectiveOut[];
}